+++ demisto_sdk/__main__.py
@@ -24,6 +24,7 @@ from demisto_sdk.commands.generate_test_playbook.test_playbook_generator import 
 from demisto_sdk.commands.generate_docs.generate_integration_doc import generate_integration_doc
 from demisto_sdk.commands.generate_docs.generate_script_doc import generate_script_doc
 from demisto_sdk.commands.generate_docs.generate_playbook_doc import generate_playbook_doc
+from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
 
 # Common tools
 from demisto_sdk.commands.common.tools import print_error, print_warning, get_last_remote_release_version, find_type
@@ -46,26 +47,20 @@ pass_config = click.make_pass_decorator(DemistoSDK, ensure=True)
     '-h', '--help'
 )
 @click.option(
-    '-d', '--env-dir', help='Specify a working directory.'
-)
-@click.option(
     '-v', '--version', help='Get the demisto-sdk version.',
     is_flag=True, default=False, show_default=True
 )
 @pass_config
-def main(config, version, env_dir):
+def main(config, version):
     config.configuration = Configuration()
     cur_version = get_distribution('demisto-sdk').version
     last_release = get_last_remote_release_version()
     if last_release and cur_version != last_release:
         print_warning(f'You are using demisto-sdk {cur_version}, however version {last_release} is available.\n'
-                      f'You should consider upgrading via "pip install --upgrade demisto-sdk" command.')
+                      f'You should consider upgrading via "pip3 install --upgrade demisto-sdk" command.')
     if version:
         version = get_distribution('demisto-sdk').version
         print(version)
-
-    if env_dir:
-        config.configuration.env_dir = env_dir
 
 
 # ====================== extract ====================== #
@@ -178,9 +173,6 @@ def unify(**kwargs):
 @click.option(
     '-j', '--conf-json', is_flag=True,
     default=False, show_default=True, help='Validate the conf.json file.')
-@click.option(
-    '-i', '--id-set', is_flag=True,
-    default=False, show_default=True, help='Create the id_set.json file.')
 @click.option(
     '--prev-ver', help='Previous branch or SHA1 commit to run checks against.')
 @click.option(
@@ -509,21 +501,32 @@ def init(**kwargs):
     "-i", "--input", help="Path of the yml file.", required=True)
 @click.option(
     "-o", "--output", help="The output dir to write the documentation file into,"
-                           " documentation file name is README.md.", required=True)
+                           " documentation file name is README.md. If not specified, will be in the yml dir.",
+    required=False)
 @click.option(
-    "-t", "--file_type", type=click.Choice(["integration", "script", "playbook"]),
-    help="The type of yml file.", required=False)
+    "-uc", "--use_cases", help="For integration - Top use-cases. Number the steps by '*' (i.e. '* foo. * bar.')",
+    required=False)
 @click.option(
-    "-e", "--examples", help="For integration - Path for file containing command or script examples."
+    "-e", "--examples", help="Path for file containing command or script examples."
                              " Each Command should be in a separate line."
                              " For script - the script example surrounded by double quotes.")
 @click.option(
-    "-id", "--id_set", help="Path of updated id_set.json file.", required=False)
+    "-p", "--permissions", type=click.Choice(["none", "general", "per-command"]), help="Permissions needed.",
+    required=True, default='none')
+@click.option(
+    "-cp", "--command_permissions", help="Path for file containing commands permissions"
+                                         " Each command permissions should be in a separate line."
+                                         " (i.e. '!command-name Administrator READ-WRITE')", required=False)
+@click.option(
+    "-l", "--limitations", help="Known limitations. Number the steps by '*' (i.e. '* foo. * bar.')", required=False)
+@click.option(
+    "--insecure", help="Skip certificate validation to run the commands in order to generate the docs.",
+    is_flag=True)
 @click.option(
     "-v", "--verbose", is_flag=True, help="Verbose output - mainly for debugging purposes.")
-def generate_doc(file_type, **kwargs):
-    input_path = kwargs['input']
-    output_path = kwargs['output']
+def generate_doc(**kwargs):
+    input_path = kwargs.get('input')
+    output_path = kwargs.get('output')
 
     # validate inputs
     if input_path and not os.path.isfile(input_path):
@@ -538,10 +541,12 @@ def generate_doc(file_type, **kwargs):
         print_error(F'Output directory {output_path} was not found.')
         return 1
 
-    if not file_type:
-        file_type = find_type(kwargs.get('input', ''))
+    file_type = find_type(kwargs.get('input', ''))
+    if file_type not in ["integration", "script", "playbook"]:
+        print_error(F'File is not an Integration, Script or a Playbook.')
+        return 1
 
-    print(f'Start generate {file_type} documentation...')
+    print(f'Start generating {file_type} documentation...')
     if file_type == 'integration':
         return generate_integration_doc(**kwargs)
     elif file_type == 'script':
@@ -551,6 +556,18 @@ def generate_doc(file_type, **kwargs):
     else:
         print_error(f'File type {file_type} is not supported.')
         return 1
+
+
+@main.command(name="create-id-set",
+              short_help='''Create the content dependency tree by ids.''')
+@click.help_option(
+    '-h', '--help'
+)
+@click.option(
+    "-o", "--output", help="Output file path, the default is the Tests directory.", required=False)
+def id_set_command(**kwargs):
+    id_set_creator = IDSetCreator(**kwargs)
+    id_set_creator.create_id_set()
 
 
 @main.resultcallback()