+++ demisto_sdk/commands/common/hook_validations/readme.py
@@ -1,6 +1,9 @@
-import subprocess
+from pathlib import Path
 import os
-from demisto_sdk.commands.common.tools import print_error, print_warning
+from demisto_sdk.commands.common.tools import print_error, print_warning, run_command_os, get_content_path
+
+NO_HTML = '<!-- NOT_HTML_DOC -->'
+YES_HTML = '<!-- HTML_DOC -->'
 
 
 class ReadMeValidator:
@@ -15,64 +18,70 @@ class ReadMeValidator:
             export DEMISTO_README_VALIDATION=True
     """
 
-    def __init__(self, file_path):
-        self.file_path = file_path
+    def __init__(self, file_path: str):
+        self.content_path = get_content_path()
+        self.file_path = Path(file_path)
+        self.pack_path = self.file_path.parent
+        self.node_modules_path = self.content_path / Path('node_modules')
 
-    def is_valid_file(self):
+    def is_valid_file(self) -> bool:
         """Check whether the readme file is valid or not
+        Returns:
+            bool: True if env configured else Fale.
         """
-        if os.environ.get('DEMISTO_README_VALIDATION'):
-            is_readme_valid = all([
-                self.is_mdx_file(),
-            ])
-            return is_readme_valid
+        if os.environ.get('DEMISTO_README_VALIDATION') or os.environ.get('CI'):
+            return self.is_mdx_file()
         else:
             return True
 
     def is_mdx_file(self) -> bool:
-        ready, is_valid = self.are_modules_installed_for_verify()
-        if ready:
-            mdx_parse = f'{os.path.dirname(os.path.abspath(__file__))}/../mdx-parse.js'
+        html = self.is_html_doc()
+        valid = self.are_modules_installed_for_verify()
+        if valid and not html:
+            mdx_parse = Path(__file__).parent.parent / 'mdx-parse.js'
+            # add to env var the directory of node modules
+            os.environ['NODE_PATH'] = str(self.node_modules_path) + os.pathsep + os.getenv("NODE_PATH", "")
             # run the java script mdx parse validator
-            res = subprocess.run(['node', mdx_parse, '-f', self.file_path], text=True, timeout=10,
-                                 capture_output=True)
-            if res.returncode != 0:
-                print_error(f'Failed verifying README.md, Path: {self.file_path}. Error Message is: {res.stderr}')
-                is_valid = False
-        return is_valid
+            _, stderr, is_valid = run_command_os(f'node {mdx_parse} -f {self.file_path}', cwd=self.content_path, env=os.environ)
+            if is_valid:
+                print_error(f'Failed verifying README.md, Path: {self.file_path}. Error Message is: {stderr}')
+                return False
+        return True
 
-    def are_modules_installed_for_verify(self):
-        node_modules_directory = os.path.dirname(os.path.dirname(
-            os.path.dirname(os.path.dirname(os.path.dirname(os.path.join(__file__))))))
-        is_valid = True
-        ready = False
-        try:
-            # check if requiring modules in node exist
-            is_node = subprocess.run(['node', '-v'], text=True, timeout=10, capture_output=True,
-                                     cwd=node_modules_directory)
-            is_mdx = subprocess.run(['npm', 'ls', '@mdx-js/mdx'], text=True, timeout=10, capture_output=True,
-                                    cwd=node_modules_directory)
-            is_fs_extra = subprocess.run(['npm', 'ls', 'fs-extra'], text=True, timeout=10, capture_output=True,
-                                         cwd=node_modules_directory)
-            is_commander = subprocess.run(['npm', 'ls', 'commander'], text=True, timeout=10, capture_output=True,
-                                          cwd=node_modules_directory)
-            if is_node.returncode == 0 and is_mdx.returncode == 0 and is_fs_extra.returncode == 0 and \
-                    is_commander.returncode == 0:
-                ready = True
-            else:
-                if is_mdx.returncode:
-                    print_warning(f"The npm module: @mdx-js/mdx is not installed in the "
-                                  f" directory:{node_modules_directory}, Test Skipped.")
-                if is_fs_extra.returncode:
-                    print_warning(f"The npm module: fs-extra is not installed in the "
-                                  f" directory: {node_modules_directory}, Test Skipped.")
-                if is_commander.returncode:
-                    print_warning(f"The npm module: commander is not installed in the "
-                                  f"directory: {node_modules_directory}, Test Skipped.")
-        except Exception as err:
-            if "No such file or directory: 'node': 'node'" in str(err):
-                print_warning(f'There is no node installed on the machine, Test Skipped, warning: {err}')
-            else:
-                print_error(f'Failed while verifying README.md, Path: {self.file_path}. Error Message is: {err}')
-                is_valid = False
-        return ready, is_valid
+    def are_modules_installed_for_verify(self) -> bool:
+        """ Check the following:
+            1. npm packages installed - see packs var for specific pack details.
+            2. node interperter exists.
+        Returns:
+            bool: True If all req ok else False
+        """
+        missing_module = []
+        valid = True
+        # Check node exist
+        stdout, stderr, exit_code = run_command_os('node -v', cwd=self.content_path)
+        if exit_code:
+            print_warning(f'There is no node installed on the machine, Test Skipped, error - {stderr}, {stdout}')
+            valid = False
+        else:
+            # Check npm modules exsits
+            packs = ['@mdx-js/mdx', 'fs-extra', 'commander']
+            for pack in packs:
+                stdout, stderr, exit_code = run_command_os(f'npm ls {pack}', cwd=self.content_path)
+                if exit_code:
+                    missing_module.append(pack)
+        if missing_module:
+            valid = False
+            print_warning(f"The npm modules: {missing_module} are not installed, Test Skipped, use "
+                          f"'npm install <module>' to install all required node dependencies")
+        return valid
+
+    def is_html_doc(self) -> bool:
+        txt = ''
+        with open(self.file_path, 'r') as f:
+            txt = f.read()
+        if txt.startswith(NO_HTML):
+            return False
+        if txt.startswith(YES_HTML):
+            return True
+        # use some heuristics to try to figure out if this is html
+        return txt.startswith('<p>') or ('<thead>' in txt and '<tbody>' in txt)